requirements.txt
# LangChain + LangGraph
langchain==0.3.7
langgraph==0.2.45
langchain-community==0.3.5
langchain-google-genai>=0.0.5  # Better Gemini integration with LangChain

# Google Gemini
google-generativeai==0.8.3

# ChromaDB
chromadb==0.5.23
langchain-chroma

# MCP Server
mcp>=0.9.0
mcp[cli]>=0.9.0  # CLI tools for MCP
httpx>=0.25.0  # For MCP HTTP transport
anyio>=4.0.0   # For async support in MCP

# Splunk
splunk-sdk>=2.0.0
xmltodict>=0.13.0  # For parsing Splunk XML responses

# Utils
python-dotenv==1.0.1
pydantic==2.9.2
requests==2.32.3
aiohttp>=3.9.0  # For async HTTP requests
pyyaml>=6.0.0   # For config files

# Jupyter
jupyter==1.1.1
notebook==7.2.2
ipykernel==6.29.5
ipywidgets>=8.0.0  # For interactive widgets in notebooks

# Data
pandas==2.2.3
numpy>=1.24.0,<2.0.0
matplotlib>=3.7.0  # For visualizations
seaborn>=0.12.0    # For better visualizations

# JSON/Data Processing
jq>=1.6.0
jsonschema>=4.19.0  # For validating JSON schemas

# Web UI
streamlit>=1.28.0

# Monitoring & Logging
loguru>=0.7.0      # Better logging for MCP server
tenacity>=8.2.0    # Retry logic for API calls

# Testing (optional but recommended)
pytest>=7.4.0
pytest-asyncio>=0.21.0  # For testing async MCP functions

# Type hints
typing-extensions>=4.8.0